Adding authorization lyre.

+++ controllers/user.controller.js
@@ -4,12 +4,17 @@ const asyncWrapper = require(join(__dirname, "..", "middlewares", "asyncWrapper"
 const User = require("../models/user.model");
 const { SUCCESS } = require(join(__dirname, "..", "utils", "httpRespondStatus"));
 const bcrypt = require("bcrypt");
+const { ADMIN } = require("../utils/rolesConstants");
+const { UNAUTHORIZED } = require("../utils/errorsConstants");
 const AppError = require(join(__dirname, "..", "utils", "AppError"))
 
 const getUsers = asyncWrapper(
     async (req, res, next)=>
     {
-        const users = await User.find({}, {"__v": false, "password": false});
+        const limit = req.query.limit? req.query.limit:10;
+        const page = req.query.page? req.query.page:1;
+        const skip = (page - 1) * limit;
+        const users = await User.find({}, {"__v": false, "password": false}).limit(limit).skip(skip);
         res.status(200).json({status: SUCCESS, data: {users}}).end();
     }
 );
@@ -17,6 +22,12 @@ const getUsers = asyncWrapper(
 const getUser = asyncWrapper(
     async (req, res, next)=>
     {
+        if(req.authData.role != ADMIN && req.authData.id != req.params.id)
+        {
+            const unauthorized = new AppError(UNAUTHORIZED, "you don't have permission to get user data");
+            return next(unauthorized);
+        }
+
         const userId = req.params.id;
         const user = await User.findOne({_id: userId}, {"__v": false, "password": false});
         if(!user)
@@ -34,7 +45,11 @@ const updateUser = asyncWrapper(
     {
         const { name, email, password, role } =  req.body;
         const userId = req.params.id;
-
+        if(req.authData.role != ADMIN && userId != req.authData.id)
+        {
+            const unauthorized = new AppError(UNAUTHORIZED, "you don't have permission to modify user data");
+            return next(unauthorized);
+        }
         const updateDataSchema = Joi.object({
             name: Joi.string().min(5),
             email: Joi.string().email(),
@@ -57,10 +72,10 @@ const updateUser = asyncWrapper(
         }
 
         console.log({updateUser: val});
-        // if(admin && // check the user role is not admin)
-        // {
-        //     val.admin = undefined;
-        // }
+        if(req.authData.role != ADMIN)
+        {
+            val.role = undefined;
+        }
 
         const user = await User.findByIdAndUpdate(userId, val, {"__v": false, "password": false});
         if(!user)
@@ -77,7 +92,11 @@ const deleteUser = asyncWrapper(
     async (req, res, next)=>
     {
         const userId = req.params.id;
-
+        if(req.authData.role != ADMIN && userId != req.authData.id)
+        {
+            const unauthorized = new AppError(UNAUTHORIZED, "you don't have permission to delete user");
+            return next(unauthorized);
+        }
         const user = await User.findByIdAndDelete({_id: userId}, {"__v": false, "password": false});
         if(!user)
         {
